Add ability to get skill info
* Get skill info by first getting vendor ID
* PHP for SMAPI GET request
* JSNI overlay class for skills info

NOTE/TODO: currently need to click Get Skill Info twice, as the first time,
the vendor ID is gotten, and then the second time, the vendor ID is
used. This is an issue with the callbacks. Should get the vendor ID in
LOGIN WITH AMAZON and only enable the Get Skill Info button once the
callback for vendor ID is executed.